% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correction.R
\name{hg_density}
\alias{hg_density}
\title{Determine the density of mercury}
\usage{
hg_density(temp)
}
\arguments{
\item{temp}{Temperature in Kelvin, or alternatively a variable with temperature units from the \code{units} package.}
}
\value{
The density of mercury
}
\description{
Determine the density of mercury
}
\references{
\insertRef{noauthor_recognized-value_1985}{pwrcorr}
}
