% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correction.R
\name{hg_pressure}
\alias{hg_pressure}
\title{Determine the pressure of mercury}
\usage{
hg_pressure(temp, in_Hg)
}
\arguments{
\item{temp}{Temperature in Kelvin, or alternatively a variable with temperature units from the \code{units} package.}

\item{in_Hg}{Pressure measured in inches of mercury without temperature adjustment}
}
\value{
The equivalent pressure in Pa
}
\description{
This function takes into consideration the effects of
temperature on the density of mercury
}
