% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correction.R
\name{pwr_corr}
\alias{pwr_corr}
\title{Power Correction Reference}
\usage{
pwr_corr(ref_temp, ref_press, humidity = NA, friction = NA)
}
\arguments{
\item{ref_temp}{Temperature in Kelvin, or alternatively a variable with temperature units from the \code{units} package.}

\item{ref_press}{Pressure in Pascals, or alternatively a variable with pressure units from the \code{units} package.}

\item{humidity}{The Relative Humidity stipulated in the reference conditions.}

\item{friction}{The friction losses used in the the reference conditions}
}
\value{
A function that outputs a correction factor
}
\description{
The function allows for the definition of various reference conditions based on
the reference pressure, temperature, humidity, and friction. \code{NA} in humidity means that
the barrometer pressure is not reduced by teh saturated vapour pressure. \code{NA} in friction
means that friction losses are not considered
}
