% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correction.R
\name{sat_vap_press}
\alias{sat_vap_press}
\title{Determine the Saturated Vapour Pressure of Water}
\usage{
sat_vap_press(temp)
}
\arguments{
\item{temp}{Temperature in Kelvin, or alternatively a variable with temperature units from the \code{units} package.}
}
\value{
Returns the saturated vapour pressure in Pa. If a temperature with units is entered then units are also returned in the output.
}
\description{
The formula used inside the equation is valid between 0 and 272 degrees Celcius and is from
the IAPWS standard \insertCite{wagner_iapws_2002}{pwrcorr}.
}
\references{
\insertRef{wagner_iapws_2002}{pwrcorr}
}
